nas/serveur_mcp/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.26
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.27.0
websockets==12.0
transformers==4.38.2
llama-cpp-python==0.2.26
numpy==1.26.3
sentencepiece==0.1.99
python-multipart==0.0.6
Jinja2==3.1.3
langchain==0.1.4
pyjwt==2.8.0
matplotlib==3.8.2
scipy==1.12.0
